getioncoloc affiche la liste de colocations du propriétaire

+++ src/main/java/fr/formation/afpa/dao/IUtilisateurDao.java
@@ -8,8 +8,8 @@ import fr.formation.afpa.domain.AppUser;
 public interface IUtilisateurDao extends JpaRepository<AppUser, Integer> {
 	
 	public AppUser findByUserName(String login);
-	public AppUser findByUserId(Long userId);
-	public void deleteByUserId (Long userId);
+	public AppUser findByUserId(int userId);
+	public void deleteByUserId (int userId);
 	public AppUser findByMail(String email);
 	
 }
